[Chore] Move ARC-20/ARC-22 test vectors to utilities/test
Relocate the inline ARC20_TOKEN_PROGRAM and ARC22_TOKEN_PROGRAM
fixtures and the compiled .aleo test vectors from the program.rs test
module into wasm/src/utilities/test/programs.rs, matching the existing
convention for shared test program fixtures.
