+++ wasm/src/utilities/test/compliant_token_template.aleo
