+++ wasm/src/utilities/test/test_usdcx_stablecoin.aleo
